requirements.txt
python-dotenv
requests
sendgrid==6.6.0
pandas
requests
Flask

pytest
